src/requirements.txt
numpy==1.24.2; python_version >= "3.9" and python_version < "4.0"
pandas==2.0.0; python_version >= "3.9" and python_version < "4.0"
matplotlib==3.7.1; python_version >= "3.9" and python_version < "4.0"
scikit-learn==1.2.2; python_version >= "3.9" and python_version < "4.0"
scipy==1.9.3 ; python_version >= "3.9" and python_version < "4.0"
PyYAML==6.0; python_version >= "3.9" and python_version < "4.0"
requests==2.30.0; python_version >= "3.9" and python_version < "4.0"
joblib==1.2.0; python_version >= "3.9" and python_version < "4.0"
ipython==8.12.0; python_version >= "3.9" and python_version < "4.0"
boto3==1.26.129; python_version >= "3.9" and python_version < "4.0"
botocore==1.29.129; python_version >= "3.9" and python_version < "4.0"
jmespath==1.0.1; python_version >= "3.9" and python_version < "4.0"
python-dateutil==2.8.2; python_version >= "3.9" and python_version < "4.0"
s3transfer==0.6.1; python_version >= "3.9" and python_version < "4.0"
six==1.16.0; python_version >= "3.9" and python_version < "4.0"
urllib3==1.26.15; python_version >= "3.9" and python_version < "4.0"
pytest==7.1.1; python_version >= "3.9" and python_version < "4.0"